added to do list app functionality and styling
