+++ partials/_app-content.scss
@@ -1,0 +1,52 @@
+.app-container {
+  &_calculator {
+    &--hidden {
+      display: none;
+    }
+    background-color: white;
+    position: absolute;
+    top: 3em;
+    left: 3em;
+    height: 50%;
+    width: 50%;
+    &_content {
+      border: 2px solid red;
+    }
+  }
+  &_toDoList {
+    &--hidden {
+      display: none;
+    }
+    background-color: white;
+    position: absolute;
+    top: 3em;
+    left: 8em;
+    height: 50%;
+    width: 50%;
+    &_content {
+      height: 100%;
+      display: flex;
+      flex-direction: column;
+      justify-content: flex-start;
+      align-items: center;
+      overflow-y: scroll;
+      background-color: #ebd8c1;
+      color: #01004e;
+      font-family: Arial, Helvetica, sans-serif;
+    }
+  }
+  &_rps {
+    &--hidden {
+      display: none;
+    }
+    background-color: white;
+    position: absolute;
+    top: 6em;
+    left: 14em;
+    height: 50%;
+    width: 50%;
+    &_content {
+      border: 2px solid red;
+    }
+  }
+}
