+++ styles.scss
@@ -1,5 +1,6 @@
 @import "./partials/apps.scss";
 @import "./partials/menu_open.scss";
+@import "./partials/app-content.scss";
 
 html,
 body {
